backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pyjwt==2.8.0
groq==0.9.0
python-dotenv==1.0.1
SQLAlchemy
sentence-transformers
scikit-learn